refactor: remove validator files
